dev_db/requirements.txt
mypy==1.9.0
mypy-extensions==1.0.0
psycopg2-binary==2.9.9
tomli==2.0.1
typing_extensions==4.11.0
